pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sparkctl"
version = "0.1.0"
description = "Orchestrates Spark standalone clusters on HPCs."
readme = "README.md"
requires-python = ">=3.11, <3.14"
license = "BSD-3-Clause"
keywords = []
authors = [
    { name = "Daniel Thom", email = "daniel.thom@nrel.gov" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    "click >= 8.2, < 9",
    "dynaconf",
    "loguru~=0.7.2",
    "psutil",
    "pydantic >= 2.7, < 3",
    "rich_click",
    "toml",
]

[project.optional-dependencies]
dev = [
    "mypy >=1.13, < 2",
    "myst_parser",
    "pre-commit",
    "pyspark-client==4.0.0",
    "pytest",
    "pytest-cov",
    "requests",
    "types-requests",
    "types-toml",
]

[project.urls]
Documentation = "https://github.com/NREL/sparkctl#readme"
Issues = "https://github.com/NREL/sparkctl/issues"
Source = "https://github.com/NREL/sparkctl"

[project.scripts]
sparkctl = "sparkctl.cli.sparkctl:cli"

[tool.mypy]
check_untyped_defs = true
files = [
  "src",
  "tests",
]

[tool.pytest.ini_options]
pythonpath = "src"
minversion = "6.0"
addopts = "-ra"
testpaths = ["tests"]
norecursedirs = ["tests/data"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".git",
    ".ruff_cache",
    ".venv",
    "_build",
    "build",
    "dist",
    "env",
    "venv",
]

line-length = 99
indent-width = 4

target-version = "py312"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = [
    "C901",   # McCabe complexity
    "E4",     # Subset of pycodestyle (E)
    "E7",
    "E9",
    "EM",     # string formatting in an exception message
    "F",      # Pyflakes
    "W",      # pycodestyle warnings
]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"

[dependency-groups]
dev = [
    "ipython>=9.1.0",
    "types-python-dateutil>=2.9.0.20241206",
    "mypy >=1.13, < 2",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"**/{tests,docs,tools}/*" = ["E402"]
